Week12_Wed

+++ Week12/SizeT.cpp
@@ -5,8 +5,8 @@ using namespace std;
 int main() {
     vector<double> values = {1.5, 2.0, 3.5};
 
-    for (int i = 0; i < values.size(); i++) {  // warning here
-    // for (size_t i = 0; i < values.size(); i++) {
+    // for (int i = 0; i < values.size(); i++) {  // warning here
+    for (size_t i = 0; i < values.size(); i++) {
         values[i] *= 2;
     }
